OWT-476: Fixed bug which was keeping old searches.
Fixed a bug in the OpenWIS user portal which was merging the
parameters of previous searches with the paramaters of brand new
ones, resulting in a more restrictive subset of results than
expected. This history is used for paging so added a parameter which
will be set when conducting new searches to bypass this.