OWT-481: Deleting bulk metadata speedup
Created branch to work on improving the performance of deleting
metadata in bulk. Added a service method on ProductMetadataService
which accepts a list of URNs (the method uses the existing metadata
deletion method at the moment). Also made changes in the user portal
to use the service method, and the bulk index deletion method when
cleaning a harvest task.