Improvement: Administration Portal - Monitoring Content - Ability to apply search criteria on a specific category
In order to monitor the content of the catalogue, the administrator can
enter a text and select the matching field (one of the possible field is
the category). 
Then, it is not possible to find metadata from a category with another
matching field. 

When the catalogue is organized in several categories, it is more
confortable to select the category first and then to apply the search on
another field (title, abstract or urn for instance).


Improvement: OWT-485