Fixed the Request/Subscribe buttons for non-GTS MD
The request and subscribe button for non-global metadata that is backed
by a local data source was not displaying the request/subscription
sub-selection parameter box.  Instead, a blank browser window was being
opened.  This update fixes this so that if no download URL is defined,
it assumes that the data is retrievable from a local data source.