Issue 48: Updating Subscriptions
Fixed a bug which was preventing users from updating geographic-based
subselection parameters for subscriptions.  This involved changing the
DTO model of the sub-selection parameters.  Previously, parameter values
were singletons (strings), which differed to how they are represented in
the database and incoming JSON models (arrays).  This has been changed
on SubSelectionParametersDTO, AdhocDTO and SubscriptionDTO.

This also involved changes to the JavaScript layer which was previously
splitting values on the commas (the original source of the bug).  The
JavaScript layer now treats these parameter values as arrays.