Issue 63: Regression in GISC-DCPC redirection
Fixed the logic used to determine whether to display a browser window
or the sub-selection dialog for DCPC metadata.  Originally, it made that
decision based on whether the MD references a localDataSource or not.
Now, a browser window is displayed if the MD references a URL which
has either "/retrieve/request/" or "/retrieve/subscribe/" somewhere in
the path (see schema/iso19139/solr-fields.xsl).