Renamed the persistence unit and fixed POM dependency
There were two persistence units being deployed and JBoss was getting
confused about which one to use.  Removed one which was not specifying
anything and renamed the persistence unit which contained all the
configuration for JPA.

Also removed the dependency for the EJBs in the cache WAR as that was
deploying the UserAlarmManager twice.