Cleaned up direct access to EJBs via JNDI
The use of InitialContext to retrieve beans, using various means of
storing the EJB's name using properties and hard-coded constants, have
been cleaned up.  Most instances of retrieving beans via the
InitialContext, particularly if the JNDI name has been hard-coded, have
been changed to make use of the facade pattern, with a single class
providing access to these beans.  This centralises the location of bean
JNDI names, making it easier to change should we need to do so in the
future (e.g. when moving to wildfly).  Instances that lookup the bean's
JNDI name from configuration remain in place.