Fixed the namespace of the UserAlarms web-service
The namespace of the UserAlarms web-service is now the same as all the
other data service web-services (excluding the cache).  This resulted in
a different package for the generated class files, which required
changes to the metadata portal.

Also moved the migration scripts for the data services into a
separate sub-directory to separate them from start/stop scripts.