Issue 92: Removed two dependencies provided by Jboss
Removed the jboss-as-client dependency from
openwis-dataservice-common-domain and changed the scope of the
jboss-ejb3-ext-api from compile to provided in
openwis-dataservice-cache-ejb.  Have managed to the data services
deploying again in JBoss AS 5.1

Note: the original stacktrace of this commit was the "Wrong target.
org.jboss.jpa.deployment.PersistenceUnitDeployment" error.  However, the
stacktrace in this issue was also encountered and, I believe, should be
fixed with this commit.