Property files is now read directoy from the file system
Have modified the ConfigSource and LocalDataSourceExtractServiceImpl to
read property files, which were originally stored in JNDI, directly from
the file system.  This removes the need to deploy the custom module and
adding system properties or configuring JNDI.

Currently the path is hard-coded to ~/conf and does not handle property
override.