Implemented proper loading of configuration
Implemented the proper loading and priority order of properties file for
the data and management service.  The defaults for both
openwis-dataservice.properties and localdatasourceservice.properties are
now embedded in the openwis-management-service-common project.
Property files overriding selected properties can be deployed to the
file system.  The location of the configuration directory can be set
with the system property "openwis.dataService.configDir" and defaults to
"~/conf".