Re-engineered protection against changes to user profile
Previous the existing user was loaded and the updates made to specific
fields.  This was causing NPE in the security services and was changed
to the update technique used in the original "save" service, but with
explit nulling of fields which could be used to change the user profile
setting.