issue-169: fixed dependency issues in load-test-data
load-test-data was previously using slf4j for logging, which was not
declared in the POM file.  Removed that dependency from that project.

Also fixed a config bug which was causing Maven build issues for the
load-test-data project due to the inherited configuration of the
exec:exec plugin.  The plugin was executing during the install goal,
attempting to install a j2ee dependency which was not present in the
project.