issue-189: fixed the browse cache content filter
Fixed a bug in the browse cache content page which was not maintaining
the set metadata filter when paging through issues.  The cause of this
was a bad request parameter.

Also fixed a bug in the Vagrant portal provisioning script which was not
properly configuring the portals for remote debugging.  Also added the
.vagrant file to .gitignore.