issue-192: added missing entities to the data service persistence.xml
Not quite sure why the persistence.xml was working previously.  It might
be because the file was originally within the domain project, and moving
into the ejb project exposed it to other entities that may not have been
automatically registered.  In any case, by adding these entities, the
data services deploy successfully and pass the automated tests.