Reconfig of vagrant to disable swapping of SSH keys
Reconfigured Vagrant to disable the swapping of the public insecure SSH
key with a generated secure one.  This is due to some failures to
authenticate in time on some host machines, causing Vagrant to abandon
provisioning the VMs.  Since these VMs are development only, this, to
me, is considered safe.