replace the use of ds-pwp-password-expiration-time
This attribute, according to OpenDJ doc, keep increasing even if the account has been blocked. Therefore, the use of this attribute is useless because we want to show a the date when the password expired (which is fixed).
To do that, we just compute the pws expired date by shifting the pwdChangedTime with the period of expiration