check request method and content type
new attribute have been added for a service: methods and contentType which defines the allowed method and the contentType. By default, get and post are allowed and all types.
Jeeves checks, before dispatching the service, that the request method and content type matches the service defined method and content type. if not, it raises an exceptions otherwise the service is dispatched.