Bug fixed when a candidate can retrieve credentials after a recover account
There is some problem with the user account. When they requested for an account, they receive a notification email which is correct. Then their account need to be elevated from candidate to user by the admin.
But when their privilege has not been elevated from candidate to user, some of them tried to request for lost account recovery and they could receive a login credential this way. I think candidate should not receive login credential if their account is not yet registered (elevated to user or admin). I have tried myself and I also face the same issue.